No longer store refs to indices and vertices
OpenGL copies vertex and index info the graphics Card. This should have
been obvious in hindsight, the Card has VRAM for a reason.
Not storing refs means we need to store properties about the data,
manually rather than getting it from the ref, but it allows for way
simpler lifetimes